+++ include/seqan3/range/views/all.hpp
@@ -13,11 +13,11 @@
 #pragma once
 
 #include <seqan3/alphabet/views/char_to.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 #include <seqan3/alphabet/views/rank_to.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/alphabet/views/to_rank.hpp>
 #include <seqan3/alphabet/views/trim_quality.hpp>
-#include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/enforce_random_access.hpp>
 #include <seqan3/range/views/get.hpp>
 #include <seqan3/range/views/interleave.hpp>
