+++ test/include/seqan3/test/performance/naive_minimiser_hash.hpp
@@ -17,8 +17,8 @@
 #include <range/v3/view/sliding.hpp>
 
 #include <seqan3/alphabet/concept.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/search/views/minimiser.hpp>
