+++ test/snippet/alphabet/views/complement.cpp
@@ -1,6 +1,7 @@
-#include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/range/views/complement.hpp>
 #include <seqan3/std/ranges>
+
+#include <seqan3/alphabet/nucleotide/dna5.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 
 int main()
 {
