+++ test/snippet/range/views/range_view_all_composability.cpp
@@ -1,6 +1,7 @@
-#include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/complement.hpp>
 #include <seqan3/std/ranges>
+
+#include <seqan3/alphabet/nucleotide/dna4.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 
 int main()
 {
