+++ test/snippet/range/views/range_view_all_notation.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/complement.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 
 int main()
 {
