+++ test/snippet/range/views/translate_usage.cpp
@@ -1,5 +1,5 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/range/views/complement.hpp>
+#include <seqan3/alphabet/views/complement.hpp>
 #include <seqan3/range/views/translate.hpp>
 
 int main()
