+++ test/unit/range/views/CMakeLists.txt
@@ -3,7 +3,6 @@ add_subdirectories()
 seqan3_test(adaptor_base_test.cpp)
 seqan3_test(as_const_test.cpp)
 seqan3_test(char_to_test.cpp)
-seqan3_test(complement_test.cpp)
 seqan3_test(convert_test.cpp)
 seqan3_test(deep_test.cpp)
 seqan3_test(enforce_random_access_test.cpp)
